Don't print to the console for waitUntil()
Testing Alarm class requires some printing in the console, but those
printing clutters up the console when doing tasks in part 2.

Now, the println commands are swapped with Lib.debug commands. It will
print to the console only when the debug flag is enabled (which in the
case of Alarm class is 'w'). We need to remember to enable the flag
before testing Alarm class and disable it otherwise.